Started working on new execution engine

+++ src/lib/fvn.c
@@ -10,7 +10,7 @@ mocha_string_hash mocha_fnv_hash(const char* str)
 	const size_t length = tyran_strlen(str) + 1;
 	mocha_string_hash hash = OFFSET_BASIS;
 	for (size_t i = 0; i < length; ++i) {
-		hash ^= *str++;
+		hash ^= (uint8_t) *str++;
 		hash *= FNV_PRIME;
 	}
 	return hash;
